GenericQueue: extract add_job()
